+++ packages/riverpod_generator/integration/build_yaml/test/build_yaml_test.dart
@@ -1,4 +1,5 @@
 import 'package:build_verify/build_verify.dart';
+import 'package:build_yaml/dependencies.dart';
 import 'package:build_yaml/main.dart';
 import 'package:test/test.dart';
 
@@ -42,6 +43,27 @@ void main() {
     expect(
       myFamilyCountStreamNotifier2ProviderFamily.name,
       'myFamilyCountStreamNotifier2ProviderFamily',
+    );
+  });
+
+  test('dependencies', () {
+    expect(
+      myFamilyCalc2ProviderFamily.dependencies,
+      unorderedEquals([
+        myCountPod,
+        myCountFuturePod,
+        myCountStreamPod,
+        myCountNotifierPod,
+        myCountAsyncNotifierPod,
+        myCountStreamNotifierPod,
+        // Family
+        myFamilyCount2ProviderFamily,
+        myFamilyCountFuture2ProviderFamily,
+        myFamilyCountStream2ProviderFamily,
+        myFamilyCountNotifier2ProviderFamily,
+        myFamilyCountAsyncNotifier2ProviderFamily,
+        myFamilyCountStreamNotifier2ProviderFamily,
+      ]),
     );
   });
 }