+++ packages/riverpod_generator/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased patch
+
+- `@Riverpod(dependencies: [...])` now respects `build.yaml` options
+
 ## 2.4.3 - 2024-08-15
 
 - `riverpod_analyzer_utils` upgraded to `0.5.4`